# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/cats": {
      "post": {
        "description": "Creates a new cat record in the database. The name must be unique to avoid identification errors.",
        "operationId": "CatsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCatDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The cat has been successfully registered."
          },
          "400": {
            "description": "Invalid input data."
          },
          "401": {
            "description": "Not authorized: No token provided or token invalid."
          },
          "409": {
            "description": "Conflict: A cat with this name already exists."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register a new cat",
        "tags": [
          "Cats Management"
        ]
      },
      "get": {
        "description": "Returns a list of all cats. Supports filtering by breed, adoption status and age (kittens).",
        "operationId": "CatsController_findAll",
        "parameters": [
          {
            "name": "breed",
            "required": false,
            "in": "query",
            "description": "Filter cats by specific breed",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAdopted",
            "required": false,
            "in": "query",
            "description": "Filter by adoption status (true/false)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isKitten",
            "required": false,
            "in": "query",
            "description": "Filter only cats younger than 1 year",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "summary": "Search shelter database",
        "tags": [
          "Cats Management"
        ]
      }
    },
    "/cats/{id}": {
      "patch": {
        "description": "Updates specific fields like name, breed, or history. System fields (isAdopted) should not be modified here.",
        "operationId": "CatsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unique numerical ID of the cat",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCatDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cat info updated successfully."
          },
          "400": {
            "description": "Invalid input or ID"
          },
          "401": {
            "description": "Not authorized: No token provided or token invalid."
          },
          "404": {
            "description": "Cat not found."
          },
          "409": {
            "description": "New name already taken"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update cat information",
        "tags": [
          "Cats Management"
        ]
      },
      "get": {
        "description": "Returns full information about a specific cat, including history and owner details if adopted.",
        "operationId": "CatsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Unique numerical ID of the cat",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cat data retrieved successfully."
          },
          "400": {
            "description": "Invalid ID format. Expected an integer."
          },
          "404": {
            "description": "Cat not found."
          }
        },
        "summary": "Get cat details",
        "tags": [
          "Cats Management"
        ]
      },
      "delete": {
        "description": "Permanently deletes a cat record from the database.",
        "operationId": "CatsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Record deleted successfully."
          },
          "400": {
            "description": "Invalid ID format. Expected an integer."
          },
          "401": {
            "description": "Not authorized: No token provided or token invalid."
          },
          "403": {
            "description": "Forbidden: You do not have administrator rights."
          },
          "404": {
            "description": "Cat not found, cannot delete."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove cat from system",
        "tags": [
          "Cats Management"
        ]
      }
    },
    "/cats/{id}/adopt": {
      "patch": {
        "description": "Links a cat to a specific user. Sets adoption status to true and records the current timestamp.",
        "operationId": "CatsController_adopt",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "number",
                    "description": "Internal ID of the adopter"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Adoption process completed."
          },
          "400": {
            "description": "Cat is already adopted or invalid user ID."
          },
          "401": {
            "description": "Not authorized: No token provided or token invalid."
          },
          "404": {
            "description": "Cat or User not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Process adoption",
        "tags": [
          "Cats Management"
        ]
      }
    },
...
    "/cats/{id}/chip": {
      "post": {
        "operationId": "CatsController_chipCat",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Chipping successful, code received."
          },
          "400": {
            "description": "The cat is already microchipped or the data is incorrect."
          },
          "401": {
            "description": "Not authorized: No token provided or token invalid."
          },
          "403": {
            "description": "Forbidden: You do not have administrator rights."
          },
          "404": {
            "description": "Cat not found."
          },
          "502": {
            "description": "The external system is unavailable."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register cat with RosKotMonitoring",
        "tags": [
          "Cats Management"
        ]
      }
    },
...
    "/external-api/roskot/register-chip": {
      "post": {
        "operationId": "RosKotPermController_registerCat",
        "parameters": [
          {
            "name": "x-api-key",
            "required": true,
            "in": "header",
            "description": "Secret access key",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RosKotRegistrationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Simulating an external chip registration API",
        "tags": [
          "Perm-External-Agency (РосКотМониторинг)"
        ]
      }
    }
  },
  "info": {
    "title": "Приют для кошек",
    "description": "Система управления кошками и пользователями",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCatDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Barsik",
            "description": "The official name of the cat for registration",
            "minLength": 2
          },
          "age": {
            "type": "number",
            "example": 2,
            "description": "Age of the cat in years",
            "minimum": 0
          },
          "breed": {
            "type": "string",
            "example": "Siamese",
            "description": "Breed or closest phenotype"
          },
          "history": {
            "type": "string",
            "example": "Found in a park during winter 2024",
            "description": "Background story of how the cat arrived at the shelter"
          },
          "description": {
            "type": "string",
            "example": "A very friendly and playful cat, loves children.",
            "description": "Detailed characteristic or personality description"
          }
        },
        "required": [
          "name",
          "age",
          "breed"
        ]
      },
      "UpdateCatDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Barsik"
          },
          "age": {
            "type": "number",
            "example": 3
          },
          "breed": {
            "type": "string",
            "example": "Siamese"
          },
          "history": {
            "type": "string",
            "example": "Healthy and active"
          },
          "description": {
            "type": "string",
            "example": "Loves playing with balls"
          }
        }
      },
...
      "RosKotRegistrationDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Барсик",
            "description": "Имя кошки"
          },
          "breed": {
            "type": "string",
            "example": "Метис",
            "description": "Порода кошки"
          }
        },
        "required": [
          "name",
          "breed"
        ]
      }
    }
  }
}